5-Ignite-NodeJS-Chapter-IV-testes-e-regras-de-negocio/src/modules/accounts/useCases/authenticateUser/AuthenticateUserUseCase.ts
import { inject, injectable } from "tsyringe";
import { IUsersRepository } from "@modules/accounts/repositories/IUsersRepository";
import { compare } from "bcrypt";
import { AppError } from "@shared/errors/AppError";
import { sign } from 'jsonwebtoken'

interface IRequest {
  email: string,
  password: string
}

interface IResponse {
  user: {
    name: string,
    email: string;
  };
  token: string;
}

@injectable()
class AuthenticateUserUseCase {


  constructor(
    @inject("UserRepository")
    private userRepository: IUsersRepository
  ) { }

  async execute({ email, password }: IRequest): Promise<IResponse> {
    //Valida se usuario existe
    const user = await this.userRepository.findByEmail(email);

    if (!user) {
      throw new AppError('Email or password incorrect');
    }
    //Se senha esta correta
    const passwordMatch = await compare(password, user.password);
    if (!passwordMatch) {
      throw new AppError('Email or password incorrect');
    }

    //Gerar JsonWebToken, add informações no payload.
    const token = sign({}, 'senhatoken123456', {
      subject: user.id,
      expiresIn: '1d'
    });

    const tokenReturn: IResponse = {
      token,
      user: {
        name: user.name,
        email: user.email
      }
    }

    return tokenReturn;
  }
}

export { AuthenticateUserUseCase }